Fix firestarter-init script failing on Windows
For some reason, running vagrant ssh -c "<cmd>" on certain Windows
computers results in Vagrant asking for a password.

This fix uses an alternative syntax to run commands in the vagrant
guest. Running it this way however, results in the command being run in
an environment where there's no login shell.

Hence the manual addition of "source ~/.bash_profile" in the command to
manually set the environment variables.